GH-38 Alterando os tipos dos campos fk_observadores_id e fk_statuses_id para unsignedBigInteger

+++ database/migrations/2019_10_12_145814_create_chamados_table.php
@@ -16,8 +16,8 @@ class CreateChamadosTable extends Migration
         Schema::create('chamados', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('descricao', 255)->nullable();//nullable indica que o campo é opcional.
-            $table->integer('fk_observadores_id');
-            $table->integer('fk_statuses_id');
+            $table->unsignedBigInteger('fk_observadores_id');
+            $table->unsignedBigInteger('fk_statuses_id');
             $table->timestamps();
             $table->foreign('fk_observadores_id')->references('id')->on('observadores');
             $table->foreign('fk_statuses_id')->references('id')->on('statuses');
